+++ src/app/app.component.spec.ts
@@ -5,9 +5,8 @@ import { FirebaseTestingModule } from './firebase-testing.module';
 describe('AppComponent', () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-    imports: [FirebaseTestingModule],
+    imports: [FirebaseTestingModule, AppComponent],
     providers: [],
-    declarations: [AppComponent],
 }).compileComponents();
   });
