+++ src/app/app.component.ts
@@ -1,12 +1,32 @@
 import { Component } from '@angular/core';
 import { Auth, onAuthStateChanged, signOut, User } from '@angular/fire/auth';
-import { Router } from '@angular/router';
+import { Router, RouterLink, RouterOutlet } from '@angular/router';
 import { environment } from 'src/environments/environment';
+import { MatNavList, MatListItem } from '@angular/material/list';
+import { MatSidenavContainer, MatSidenav, MatSidenavContent } from '@angular/material/sidenav';
+import { MatIconButton } from '@angular/material/button';
+import { NgIf } from '@angular/common';
+import { MatIcon } from '@angular/material/icon';
+import { MatToolbar } from '@angular/material/toolbar';
 
 @Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss'],
+    selector: 'app-root',
+    templateUrl: './app.component.html',
+    styleUrls: ['./app.component.scss'],
+    standalone: true,
+    imports: [
+        MatToolbar,
+        MatIcon,
+        NgIf,
+        MatIconButton,
+        MatSidenavContainer,
+        MatSidenav,
+        MatNavList,
+        MatListItem,
+        RouterLink,
+        MatSidenavContent,
+        RouterOutlet,
+    ],
 })
 export class AppComponent {
   user: User | null = null;
