+++ src/main.ts
@@ -1,13 +1,99 @@
-import { enableProdMode } from '@angular/core';
-import { platformBrowser } from '@angular/platform-browser';
+import { enableProdMode, importProvidersFrom } from '@angular/core';
+import { BrowserModule, bootstrapApplication } from '@angular/platform-browser';
 
-import { AppModule } from './app/app.module';
+import { initializeApp, provideFirebaseApp } from '@angular/fire/app';
+import { connectAuthEmulator, getAuth, provideAuth } from '@angular/fire/auth';
+import {
+  connectFirestoreEmulator,
+  getFirestore,
+  provideFirestore,
+} from '@angular/fire/firestore';
+import { MatButtonModule } from '@angular/material/button';
+import { MAT_FORM_FIELD_DEFAULT_OPTIONS } from '@angular/material/form-field';
+import { MatIconModule } from '@angular/material/icon';
+import { MatListModule } from '@angular/material/list';
+import { MatSidenavModule } from '@angular/material/sidenav';
+import { MAT_SNACK_BAR_DEFAULT_OPTIONS } from '@angular/material/snack-bar';
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { provideAnimations } from '@angular/platform-browser/animations';
+import { TitleStrategy } from '@angular/router';
+import { ServiceWorkerModule } from '@angular/service-worker';
+import {
+  provideNgxWebstorage,
+  withLocalStorage,
+  withNgxWebstorageConfig,
+  withSessionStorage,
+} from 'ngx-webstorage';
+import { AppRoutingModule } from './app/app-routing.module';
+import { AppComponent } from './app/app.component';
+import { PageTitleStrategy } from './app/strategies/title-strategy';
 import { environment } from './environments/environment';
 
 if (environment.production) {
   enableProdMode();
 }
 
-platformBrowser()
-  .bootstrapModule(AppModule)
-  .catch((err) => console.error(err));
+bootstrapApplication(AppComponent, {
+  providers: [
+    importProvidersFrom(
+      BrowserModule,
+      AppRoutingModule,
+      ServiceWorkerModule.register('ngsw-worker.js', {
+        enabled: environment.production,
+      }),
+      MatButtonModule,
+      MatIconModule,
+      MatListModule,
+      MatSidenavModule,
+      MatToolbarModule,
+    ),
+    provideNgxWebstorage(
+      withNgxWebstorageConfig({
+        prefix: 'fukagawa-coffee',
+        separator: '.',
+        caseSensitive: true,
+      }),
+      withLocalStorage(),
+      withSessionStorage(),
+    ),
+    provideFirebaseApp(() => initializeApp(environment.firebase)),
+    provideFirestore(() => {
+      const fire = getFirestore();
+      if (
+        !environment.production &&
+        environment.firebaseEmulator.firestorePort
+      ) {
+        connectFirestoreEmulator(
+          fire,
+          'localhost',
+          environment.firebaseEmulator.firestorePort,
+        );
+      }
+      return fire;
+    }),
+    provideAuth(() => {
+      const auth = getAuth();
+      if (!environment.production && environment.firebaseEmulator.authUrl) {
+        connectAuthEmulator(auth, environment.firebaseEmulator.authUrl);
+      }
+      return auth;
+    }),
+    {
+      provide: MAT_FORM_FIELD_DEFAULT_OPTIONS,
+      useValue: {
+        appearance: 'fill',
+      },
+    },
+    {
+      provide: MAT_SNACK_BAR_DEFAULT_OPTIONS,
+      useValue: {
+        duration: 3000,
+      },
+    },
+    {
+      provide: TitleStrategy,
+      useClass: PageTitleStrategy,
+    },
+    provideAnimations(),
+  ],
+}).catch((err) => console.error(err));
